adabot/requirements.txt
# SPDX-FileCopyrightText: 2017 Scott Shawcroft for Adafruit Industries
#
# SPDX-License-Identifier: MIT

black==22.3.0
packaging==20.3
pylint
pytest
pyyaml==5.4.1
redis==2.10.6
requests==2.25.1
sh==1.12.14
requests-cache==0.5.2
parse==1.19.0
GitPython==3.1.27
PyGithub==1.55
